Added proper handler for report_dir creation
When concurrent tests are running, there's a race condition – if the
directory is created between the os.path.exists and the os.makedirs
calls, the os.makedirs will fail with an OSError. Unfortunately,
blanket-catching OSError and continuing is not foolproof, as it will
ignore a failure to create the directory due to other factors, such as
insufficient permissions, full disk, etc. This adds a proper way of
creating a directory, that should work both on 2.7 and 3.3+ versions
of Python.